add alert window to delete

+++ client/src/pages/LauncherDetailsPage.jsx
@@ -29,6 +29,7 @@ function LauncherDetailsPage() {
     }, [id])
 
     const handleDelete = async ()=>{
+        if(!window.confirm("You really need to delete this Launcher?")) return
         try {
             await deleteLauncher(id)
             navigate('/')
